[refactor] 로직 리팩터링
