+++ src/main/java/com/example/bolmalre/member/service/OAuthServiceImpl.java
@@ -1,6 +1,5 @@
-package com.example.bolmalre.member.oauth;
+package com.example.bolmalre.member.service;
 
-import com.example.bolmalre.auth.filter.LoginFilter;
 import com.example.bolmalre.auth.jwt.JWTUtilImpl;
 import com.example.bolmalre.auth.service.RefreshTokenService;
 import com.example.bolmalre.config.JWTConfig;
@@ -9,24 +8,19 @@ import com.example.bolmalre.member.domain.Member;
 import com.example.bolmalre.member.domain.enums.Role;
 import com.example.bolmalre.member.infrastructure.MemberRepository;
 import com.example.bolmalre.member.infrastructure.UuidHolder;
+import com.example.bolmalre.member.util.KakaoUtil;
+import com.example.bolmalre.member.web.dto.KakaoDTO;
+import com.example.bolmalre.member.web.port.OAuthService;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpHeaders;
-import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-
-import java.util.Collection;
-import java.util.Iterator;
-import java.util.Optional;
-import java.util.UUID;
 
 import static com.example.bolmalre.member.util.CookieUtil.createCookie;
 
 @Service
 @RequiredArgsConstructor
-public class AuthService {
+public class OAuthServiceImpl implements OAuthService {
 
     private final KakaoUtil kakaoUtil;
     private final MemberRepository memberRepository;
@@ -38,7 +32,8 @@ public class AuthService {
 
     private final UuidHolder uuid;
 
-
+    // 추후 추가 정보 기입을 위한 로직 필요
+    @Override
     public Member oAuthLogin(String accessCode, HttpServletResponse httpServletResponse) {
 
         KakaoDTO.OAuthToken oAuthToken = kakaoUtil.requestToken(accessCode);
@@ -48,18 +43,7 @@ public class AuthService {
         Member byEmail = memberRepository.findByEmail(requestEmail)
                 .orElseGet(() -> createNewUser(kakaoProfile));
 
-        String username = byEmail.getUsername();
-        Role role = byEmail.getRole();
-
-        //토큰 생성
-        String access = jwtUtil.createJwt("access", username, role.toString(), jwtConfig.getAccessTokenValidityInSeconds());// 1시간
-        String refresh = jwtUtil.createJwt("refresh", username, role.toString(), jwtConfig.getRefreshTokenValidityInSeconds());//30일
-
-        //응답 설정
-        httpServletResponse.setHeader("access", access);
-        httpServletResponse.addCookie(createCookie("refresh", refresh));
-
-        refreshTokenService.addRefreshEntity(username,refresh, jwtConfig.getRefreshTokenValidityInSeconds());
+        loginProcess(httpServletResponse, byEmail);
 
         return byEmail;
     }
@@ -73,5 +57,20 @@ public class AuthService {
                 uuid
         );
         return memberRepository.save(newMember);
+    }
+
+    private void loginProcess(HttpServletResponse httpServletResponse, Member byEmail) {
+        String username = byEmail.getUsername();
+        Role role = byEmail.getRole();
+
+        //토큰 생성
+        String access = jwtUtil.createJwt("access", username, role.toString(), jwtConfig.getAccessTokenValidityInSeconds());// 1시간
+        String refresh = jwtUtil.createJwt("refresh", username, role.toString(), jwtConfig.getRefreshTokenValidityInSeconds());//30일
+
+        //응답 설정
+        httpServletResponse.setHeader("access", access);
+        httpServletResponse.addCookie(createCookie("refresh", refresh));
+
+        refreshTokenService.addRefreshEntity(username,refresh, jwtConfig.getRefreshTokenValidityInSeconds());
     }
 }
