+++ src/main/java/com/example/bolmalre/member/web/port/OAuthService.java
@@ -1,0 +1,8 @@
+package com.example.bolmalre.member.web.port;
+
+import com.example.bolmalre.member.domain.Member;
+import jakarta.servlet.http.HttpServletResponse;
+
+public interface OAuthService {
+    Member oAuthLogin(String accessCode, HttpServletResponse httpServletResponse);
+}
